refactoring done

+++ src/main/java/com/example/restservice/CarsController.java
@@ -1,6 +1,7 @@
 package com.example.restservice;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,6 +11,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -47,18 +50,34 @@ public class CarsController {
   // post
   @PostMapping
   public void addNewCar(@RequestBody Car car) {
+    String makeCar = car.getMake();
+    String modelCar = car.getModel();
+    String colourCar = car.getColour();
+    Integer yearCar = car.getYear();
+    if (makeCar == null || modelCar == null || colourCar == null || yearCar == null) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "bad request. Mandatory car attributes are missing");
+    }
     carService.addNewCar(car);
   }
 
   // delete
   @DeleteMapping("{carId}")
-  public void deleteCar(@PathVariable("carId") Long carId) {
+  public void deleteCar(@PathVariable("carId") Long carId) throws ResponseStatusException {
+    Optional<Car> exists = carService.getCarById(carId);
+    if (exists.isEmpty()) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Car with Id '" + carId + "' not found");
+    }
+
     carService.deleteCar(carId);
   }
 
   // put
   @PutMapping
   public void updateCar(@RequestBody Car car) {
+    Optional<Car> exists = carService.getCarById(car.getId());
+    if (exists.isEmpty()) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "car with id " + car.getId() + " does not exist");
+    }
     carService.updateCar(car);
   }
